Mon, Oct 7, 2024, 10:52 AM -03:00

+++ week04/section/02-pdf.c
@@ -6,6 +6,6 @@ int main(int argc, string argv[])
 {
     string filename = argv[1];
     FILE *f = fopen(filename, "r");
-    int buffer[4];
-    
+    uint8_t buffer[4];
+
 }
